backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
firebase-admin==6.2.0
PyJWT==2.8.0
google-cloud-storage==2.10.0
supabase==2.0.3
pinecone-client==2.2.4
openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
langchain==0.0.350